+++ api/src/main/java/org/svip/api/controller/SVIPApiController.java
@@ -43,6 +43,7 @@ import org.svip.sbomvex.model.VEXType;
 import org.svip.sbomvex.vexstatement.VEXStatement;
 import org.svip.utils.VirtualPath;
 
+import java.io.FileNotFoundException;
 import java.util.*;
 import java.util.Arrays;
 
@@ -50,6 +51,7 @@ import java.io.IOException;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Optional;
+import java.util.zip.ZipFile;
 
 /**
  * Spring API Controller for handling requests to the SVIP backend.
@@ -138,7 +140,7 @@ public class SVIPApiController {
         } catch (IllegalArgumentException | JsonProcessingException e) {
             LOGGER.info("POST /svip/sboms - " + errorMsg);
             LOGGER.error(e.getMessage());
-            return new ResponseEntity<>(errorMsg, HttpStatus.BAD_REQUEST);
+            return new ResponseEntity<>(e.getMessage() + " " + errorMsg, HttpStatus.BAD_REQUEST);
         }
 
         // Upload
@@ -276,14 +278,14 @@ public class SVIPApiController {
      * @return converted SBOM
      */
     @PutMapping("/sboms")
-    public ResponseEntity<String> convert(@RequestParam("id") long id, @RequestParam("schema") SerializerFactory.Schema schema,
+    public ResponseEntity<Long> convert(@RequestParam("id") long id, @RequestParam("schema") SerializerFactory.Schema schema,
                                           @RequestParam("format") SerializerFactory.Format format,
                                           @RequestParam("overwrite") Boolean overwrite){
         // Get SBOM
         Optional<SBOMFile> sbomFile = sbomFileRepository.findById(id);
 
         // Check if it exists
-        ResponseEntity<String> NOT_FOUND = Utils.checkIfExists(id, sbomFile, "convert");
+        ResponseEntity<Long> NOT_FOUND = Utils.checkIfExists(id, sbomFile, "convert");
         if (NOT_FOUND != null) return NOT_FOUND;
 
         // Get and convert SBOM
@@ -318,12 +320,14 @@ public class SVIPApiController {
         converted.setFileName(toConvert.getFileName());
 
         // overwrite
-        if(overwrite){
+        if(overwrite)
             sbomFileRepository.deleteById(id);
-            sbomFileRepository.save(converted);
-        }
+        else
+            converted.setId(Utils.generateNewId(id, new Random(), sbomFileRepository));
 
-        return Utils.encodeResponse(converted.getContents());
+        sbomFileRepository.save(converted);
+
+        return Utils.encodeResponse(converted.getId());
     }
 
     /** USAGE Send GET request to /qa with a URL parameter id to conduct a quality assessment on the SBOM with
@@ -386,18 +390,16 @@ public class SVIPApiController {
      * USAGE. Send GENERATE request to /generate an SBOM from source file(s)
      *
      * @param projectName of project to be converted to SBOM
-     * @param files from project source
+     * @param zipPath path to zip file
      * @param schema to convert to
      * @param format to convert to
      * @return generated SBOM
      */
     @PostMapping("/generators/parsers")
-    public ResponseEntity<String> generateParsers(@RequestBody SBOMFile[] files,
-                                           @RequestParam("projectName") String projectName,
-                                           @RequestParam("schema") SerializerFactory.Schema schema,
-                                           @RequestParam("format") SerializerFactory.Format format){
-
-        ParserController parserController = new ParserController(projectName, new HashMap<>());
+    public ResponseEntity<?> generateParsers(@RequestParam("zipPath") String zipPath,
+                                             @RequestParam("projectName") String projectName,
+                                             @RequestParam("schema") SerializerFactory.Schema schema,
+                                             @RequestParam("format") SerializerFactory.Format format) throws IOException {
 
         String urlMsg = "GENERATE /svip/generate?projectName=" + projectName;
 
@@ -406,14 +408,26 @@ public class SVIPApiController {
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
 
-        for (SBOMFile f: files
+        ArrayList<HashMap<SBOMFile, Integer>> unZipped = (ArrayList<HashMap<SBOMFile, Integer>>) Utils.unZip(zipPath);
+
+        HashMap<VirtualPath, String> virtualPathStringHashMap = new HashMap<>();
+
+        for (HashMap<SBOMFile, Integer> h: unZipped
              ) {
+
+            SBOMFile f = (SBOMFile) h.keySet().toArray()[0];
+
             if(f.hasNullProperties()){
                 LOGGER.error(urlMsg + "/fileName=" + f.getFileName() + " has null properties");
                 return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
             }
-            parserController.addFile(new VirtualPath(f.getFileName()), f.getContents());
+
+            virtualPathStringHashMap.put(new VirtualPath(f.getFileName()), f.getContents());
+
         }
+
+        ParserController parserController = new ParserController(projectName, virtualPathStringHashMap);
+
 
         SVIPSBOM parsed;
         try{
@@ -436,12 +450,18 @@ public class SVIPApiController {
             return new ResponseEntity<>(error, HttpStatus.INTERNAL_SERVER_ERROR);
         }
 
-        return Utils.encodeResponse(contents);
+        SBOMFile result = new SBOMFile(projectName + ((format == SerializerFactory.Format.JSON)
+                 ? ".json" : ".spdx"),contents);
+        Random rand = new Random();
+        result.setId(Utils.generateNewId(rand.nextLong(), rand, sbomFileRepository));
+        sbomFileRepository.save(result);
+
+        return Utils.encodeResponse(result.getId());
 
     }
 
     @PostMapping("/generators/osi")
-    public ResponseEntity<String> generateOSI(@RequestBody SBOMFile[] files,
+    public ResponseEntity<?> generateOSI(@RequestBody SBOMFile[] files,
                                            @RequestParam("projectName") String projectName,
                                            @RequestParam("schema") SerializerFactory.Schema schema,
                                            @RequestParam("format") SerializerFactory.Format format){
@@ -457,7 +477,7 @@ public class SVIPApiController {
      * @return a merged sbomFile
      */
     @GetMapping("/merge")
-    public ResponseEntity<String> merge(@RequestParam("ids") long[] ids){
+    public ResponseEntity<?> merge(@RequestParam("ids") long[] ids){
 
         ArrayList<SBOM> sboms = new ArrayList<>();
 
@@ -473,7 +493,7 @@ public class SVIPApiController {
             Optional<SBOMFile> sbomFile = sbomFileRepository.findById(id);
 
             // Check if it exists
-            ResponseEntity<String> NOT_FOUND = Utils.checkIfExists(id, sbomFile, "merge");
+            ResponseEntity<Long> NOT_FOUND = Utils.checkIfExists(id, sbomFile, "merge");
             if (NOT_FOUND != null) return NOT_FOUND;
             SBOMFile sbom = sbomFile.get();
 
@@ -528,17 +548,12 @@ public class SVIPApiController {
         SBOMFile result = new SBOMFile(merged.getName(), contents);
         Random rand = new Random();
 
-        // assign new id and name
-        int i = 0;
-        while(sbomFileRepository.existsById(idSum)){ // todo check if frontend are okay with this
-            idSum += (rand.nextLong() + idSum) % ((i < 100) ? idSum : Long.MAX_VALUE);
-            i++;
-        }
+        idSum = Utils.generateNewId(idSum, rand, sbomFileRepository);
 
         result.setId(idSum);
         sbomFileRepository.save(result);
 
-        return Utils.encodeResponse(result.getContents());
+        return Utils.encodeResponse(idSum);
 
     }
 
@@ -638,4 +653,5 @@ public class SVIPApiController {
         return new ResponseEntity<>(
                 new VEXResult(vex, error), HttpStatus.OK);
     }
+
 }